requirements.txt
heyoo==0.0.9
python-dotenv==1.0.0
requests==2.31.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
aiohttp==3.9.1
numpy>=1.26.0
torch>=2.1.0
transformers>=4.36.0
accelerate>=0.25.0
huggingface-hub>=0.20.1
protobuf>=4.25.1
tqdm>=4.66.1
packaging>=23.2
typing-extensions>=4.9.0
regex>=2023.10.3
tokenizers>=0.15.0
